String to Int
Changed main and read data classes. Read in the data in read data class but pass in the data from main. Data is stored now as a table and all the number values have been changed from string to ints. We can now access these values and do math with them.