drawCloud() function allows chosen cloud to smoothly transition to the next page
Fixed up the button class with the new function, added switch statement in the main to determine which button was pressed and transition to chosen place. Added a way to get the x position of the clouds in order to do this. Added the new variables movedCloud, isSelected, and isMoving as well for the function.